pyproject.toml
[tool.poetry]
name = "deep-server"
version = "0.1.0"
description = ""
authors = ["deep-dev <dev@thedeep.io>"]
license = "GNU Affero General Public License v3.0"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "==4.9.3"
boto3 = "==1.17.98"
graphdoc = "0.4.0"
celery = { version = "==5.2.7", extras = ["redis"] }
# coreapi = "==2.3.3"
# dateparser = "==0.7.0"
# decorator = "==4.3.2"
# descartes = "==1.1.0"
djangorestframework-camel-case = "==1.2.0"
djangorestframework = "==3.12.4"
django-celery-beat = "==2.2.0"
django-cors-headers = "==3.7.0"
django-crispy-forms = "==1.12.0"
django-enumfield = "==2.0.2"
django-filter = "==2.4.0"
django-jsoneditor = "==0.1.6"
django-ordered-model = "==3.4.3"
django-otp = { version = "==1.0.6", extras = ["qrcode"] }
django-premailer = "==0.2.0"
django-redis = "==5.0.0"
django-reversion = "==3.0.9"
django-mptt = "==0.12.0"
django-ses = "==3.5.1"
django-storages = "==1.11.1"
django-utils-six = "2.0"  # Require by autofixture
drf-dynamic-fields = "==0.3.1"
drf-yasg = { version = "==1.20.0", extras = ["validation"] }
feedparser = "==6.0.8"
flower = { git = "https://github.com/mher/flower.git", rev = "7ee34cb5fc039bbeb208a54f113713edbb4809ce" }
google-api-python-client = "==2.9.0"
graphene-django = ">=2,<3"
graphene-graphiql-explorer = "*"
graphene-django-extras = { git = "https://github.com/the-deep/graphene-django-extras.git", branch = "v0.4.9-patch.2" }
lxml = "==5.2.2 "
mdmail = "==0.1.3"
oauth2client = "==4.1.3"
openpyxl = "==3.0.9"
parameterized = "==0.8.1"
"pdfminer.six" = "==20181108"
psycopg2-binary = "==2.9.9"
pyexcel-ods = "==0.6.0"
python-dateutil = "==2.8.1"
python-docx = "==0.8.11"
pyxform = "==1.10.1"
readability-lxml = "==0.8.1"
requests = "==2.31.0"
rest-framework-generic-relations = "==2.0.0"
sentry-sdk = "*"
tldextract = "==3.1.0"
user-agents = "==2.2.0"
Markdown = "==3.3.4"
Pillow = "==10.3.0"
PyJWT = "==2.6.0"
Pygments = "==2.9.0"
Django = ">=3.2,<3.3"  # TODO Upgrade
ipython = "*"
factory-boy = "*"
colorlog = "*"
drf-writable-nested = "*"
graphene-file-upload = "^1.3.0"
django-environ = "^0.8.1"
uwsgi = "*"
django-admin-autocomplete-filter = "^0.7.1"
aws-sns-message-validator = "0.0.5 "

# Thumbnail
# Shapely = "==1.7.1"
# pandas = "==1.2.5"
# selenium = "==3.141.0"
# Tabular needed modules
# matplotlib = "==3.4.2"
# geopandas = "==0.9.0"
# plotly = "==5.0.0"
# wordcloud = "==1.8.1"
# psutil = "==5.6.6"
datasketch = "^1.5.8"

[tool.poetry.dev-dependencies]
pytest-profiling = "*"
pytest-ordering = "*"
pytest-django = "*"
pytest-icdiff = "*"
snapshottest = "*"
django-debug-toolbar = "*"
django-graphiql-debug-toolbar = "*"
django-stubs = { version = "*", allow-prereleases = true }
django-autofixture = { branch = "master", git = "https://github.com/the-deep/django-autofixture" }
django-fixture-magic = { rev = "466fa35c075009e4a1974f6e588471dd9e720e53", git = "https://github.com/davedash/django-fixture-magic.git" }
# mypy = "*"
# flake8 = "*"
# pylint = "*"

[tool.pyright]
extraPaths = ["apps"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "apps/experimental",
    "apps/typestubs",
    "**/snap_test_*.py",
]
reportMissingImports = true
reportMissingTypeStubs = false

[tool.black]
line-length = 130
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | docs
)/
'''
# NOTE: Update in .pre-commit-config.yaml as well
extend-exclude = "^.*\\b(migrations)\\b.*$ (__pycache__|.*snap_test_.*\\.py|.+/+.+/+migrations/+.*)"

[tool.isort]
profile = "black"
multi_line_output = 3
# NOTE: Update in .pre-commit-config.yaml as well
skip = [
    "**/__pycache__",
    "**/snap_test_*.py",
    "**/migrations/*.py",
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
